split plugin and non plugin tests
